feat: solved day 19 medium

+++ Medium/Day 19/solution.cpp
@@ -4,14 +4,14 @@
 using namespace std;
 typedef long long ll;
 int main() {
-int n, x ; 
+long long int  n, x ; 
 cin >> n >> x;
-  int a[n + 1];
-  for(int i = 1; i <= n; i++) cin >> a[i];
-  int ans = 0;
-  for(int i = 2; i <= n; i++) {
+  long long int  a[n + 1];
+  for(long long int  i = 1; i <= n; i++) cin >> a[i];
+  long long int  ans = 0;
+  for(long long int  i = 2; i <= n; i++) {
     if(a[i - 1] + a[i] > x) {
-        int temp = a[i - 1] + a[i];
+        long long int  temp = a[i - 1] + a[i];
         temp = temp - x;
         ans += temp;
         if(temp > a[i]) a[i] = 0;
